+++ docs/CHANGE_LOG.md
@@ -3,7 +3,8 @@
 ## Version 2.2.0:
 
 #### **Public Functions:**
-- Added "includeDataAttributes()" support for "json()" (including binding support), which states if the data attributes should be included (defaults to true).
+- Added "includeDataAttributes()" support for "json()", which states if the data attributes should be included (defaults to true).
+- Added "addDataAttributes()" support for "html()" (including binding support), which states if the data attributes should be added (defaults to true).
 
 #### **Documentation:**
 - Updated the documentation to show the right properties for the objects, and add missing documentation for the binding attribute.